backend/app/shared/security.py
"""
Utilidades de seguridad: hashing de contraseñas, JWT y tokens de reset.
"""

import hashlib
import secrets
import uuid
from datetime import datetime, timedelta, timezone

from jose import JWTError, jwt
import bcrypt

from app.shared.config import settings

# ── Contraseñas ────────────────────────────────────────────────────────
def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Verifica una contraseña plana contra su hash usando bcrypt directo."""
    try:
        return bcrypt.checkpw(
            plain_password.encode("utf-8"), hashed_password.encode("utf-8")
        )
    except Exception:
        return False


def get_password_hash(password: str) -> str:
    """Genera un hash de bcrypt para una contraseña plana."""
    # Bcrypt maneja el salting automáticamente
    salt = bcrypt.gensalt()
    hashed = bcrypt.hashpw(password.encode("utf-8"), salt)
    return hashed.decode("utf-8")


# ── Tokens de recuperación de contraseña ───────────────────────────────
def hash_reset_token(token: str) -> str:
    return hashlib.sha256(token.encode("utf-8")).hexdigest()


def generate_reset_token() -> str:
    return secrets.token_urlsafe(32)


# ── JWT ────────────────────────────────────────────────────────────────
def create_access_token(
    *, user_id: int, role: str, expires_delta: timedelta | None = None
) -> tuple[str, str, datetime]:
    jti = str(uuid.uuid4())
    expire = datetime.now(timezone.utc) + (
        expires_delta
        if expires_delta
        else timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    )
    to_encode = {
        "sub": str(user_id),
        "role": role,
        "jti": jti,
        "exp": int(expire.timestamp()),
    }
    encoded = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
    return encoded, jti, expire


def decode_access_token(token: str) -> dict:
    return jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])


def jwt_payload_safe(token: str) -> dict | None:
    try:
        return decode_access_token(token)
    except JWTError:
        return None
