backend/requirements.txt
# ═══════════════════════════════════════════════════════════════════════
# RutAIGeoProxi — Dependencias de Producción
# Backend ultra-ligero (~50 MB) — IA procesada en la nube
# ═══════════════════════════════════════════════════════════════════════

# ── Core ──
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
email-validator>=2.0.0
websockets>=12.0
httpx>=0.27.0

# ── Base de Datos ──
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# ── Storage ──
python-multipart>=0.0.10
# firebase-admin>=6.5.0  # Opcional: solo si usas Firebase Storage (descomenta si es necesario)

# ── Reportes ──
reportlab>=4.1.0
pandas>=2.2.0
openpyxl>=3.1.0

# ── Utilidades ──
gunicorn>=21.2.0
